Upped version number to 0.3.6, and a few major changes. Those who wish to avoid any hassle may want to point their MRtrix SVN repo to https://subversion.assembla.com/svn/mrtrix_dev/tags/0.3.5 instead...
Main change relates to new command-line parsing interface, designed to allow the definition of reusable command-line options, 
so that options commonly used can simply be re-imported into other programs (e.g. "-grad"). Also allows parsing of command-line options from 
anywhere in the code, so that these commonly-used options can be parsed by the piece of code responsible for it. As part of this
new interface, options are defined as part of an option group, and listed as such in the help display. This will for example 
allow options specific to a particular tracking algorithm to define themselves as such, and also to perform their own 
command-line parsing in their own method's definition rather than in the EXECUTE method, which should help simplify maintenance 
of the code. Another use of this new interface is the possibility to have MRView's display modes and tools have their own 
command-line interface without any interference with the main application - enables better separation between the main application
"shell" and the modules (modes & tools) that plug into it.

As part of this change, the structure of MRtrix applications has also been modified. It is in my opinion simpler and more intuitive.
This commit does not include any updates to the documentation on this, but a quick look at  existing commands in the cmd folder 
should be sufficient. Updates to documentation will follow when I have a minute...