added support for no-VLA compilers
Some compilers (notably clang on MacOSX) don't allow the use of variable-length
arrays (VLA) in particular cases. In the case of clang, VLAs can't be used for
non-POD data types. Two checks were added to the configure script in a previous
commit to check for VLA and non-POD VLA support, which sets the macros
MRTRIX_NO_VLA & MRTRIX_NO_NON_POD_VLA respectively. This commit removes the use
of VLAs where doing so shouldn't hurt performance, and introduces new macros to
use VLAs if supported, or a std::vector or fixed-length array otherwise. These
macros are:

- VLA (name, type, size) : define a VLA or std::vector otherwise.
- VLA_MAX (name, type, size, max) : define a VLA or fixed-length array of size max
  otherwise

The non-POD versions are identical, but are to be used for non-POD types. This
is because the compiler might support VLAs for POD types, but not for non-POD
types, so these macros might be defined differently to the VLA version. Usage
is identical to the VLA versions:
- NON_POD_VLA (name, type, size)
- NON_POD_VLA_MAX (name, type, size, max)