MRView: separated Shader from its Displayable
Extensive changes to the way OpenGL shaders are handled. There is now a
new Shader class, which is inherited to provide the appropriate shaders
for its Displayable. These are now managed by each Mode or Tool that
needs them, rather than on a per-Displayable basis. This simplifies
switching between Modes that don't use the standard per-slice image
shader (i.e. the Volume mode). This new Shader class manages its own
recompilation as required during the paint() handler, by ensuring it is
up to date with respect to the parameters of the Displayable when it is
loaded (using the Displayable::start() function).