Help page now displayed using less
Added ability for help page of each command to be displayed using the
Unix 'less' command, with bold and underline formatting using
overstrike. This means developers can include as much information as
they want into the help page, without worrying about how big it might
end up since there is no longer any motivation to try and fit it into a
single terminal screen. It also allows the formatting of the page to be
improved and spaced out for better readability.

The command used for displaying the help page can be set in the config
file using the key 'HelpCommand', which defaults to 'less' if not found.
Setting an empty key reverts the help page to its previous behaviour.