MRView: add rudimentary support for scripting
Removed the previous option parsing infrastructure, since it was never
going to be flexible enough. Two new options have been added:

-run <command> will run a single command, e.g. mrview -run "view.mode 3"
to switch to volume mode at startup. Multiple such commands can be
supplied, and will be executed in turn.

-batch <file> will run all the commands in the specified file, one
command per line. Multiple batch files can be supplied, and will be
executed in turn. Note that all batch files currently run before any
of the commands supplied with -run.

Still need to add support for as many commands as are reasonable/needed.
In particular, need support for opening tools and having tools parse
their own commands. The plan is simply to pass any commands not
processed by the main window through to each of the currently open tools
in turn until processed, through a virtual
Tool::Base::process_batch_command() function.

In the longer run, this framework could be used to save 'session' files,
optionally with the commands required to load the images themselves.
This would require the main window and any of the open tools to dump a
set of commands that would get them back to their current state.