MRView: update all OpenGL code to use new glLoadGen cpp profile
This means that all GLenums get renamed from GL_XXX to gl::XXX, and all
functions get renamed from glFunc() to gl::Func(). Only exceptions are
GL_TRUE -> gl::TRUE_ and GL_FALSE -> gl::FALSE_ (to avoid conflicts with
macros TRUE & FALSE).

Note that the glLoadGen profile needed to change to remove all checks
for previous inclusion of OpenGL headers - the idea is to include this
*after* including the Qt OpenGL headers, which will bring the system
OpenGL headers in.  Using a namespaced glLoadGen profile should however
remove all conflicts with any other headers, so there *should* be no
reason to force it to be included first. There was however one remaining
conflict: a typedef for GLDEBUGPROC, which conflicted due to the last
argument differing ('const void*' -> 'GLvoid*') between the glLoadGen
version and the glext.h version. This is now commented out since it's
not used in any of our code.

This seems to work fine on Linux with Qt4.8. Will be checking linux with
Qt5.2 soon, and then try other platforms too.