Fixed output of streamline weights
The writing of streamline weights to file as part of the Tractography::Writer was not working, as the file was being opened with the std::ios::ate flag, which fails if the file does not exist.
Now an empty file is created on construction, which can then be appended to.