Explicit multi-thread queue handling for tckedit
When tckedit is used to truncate only a small number of streamlines, it can hang when using a batched queue; suspect this is the loader thread filling the batched queue with streamlines data that will never be used.
Therefore, use a non-batched queue if a small number of streamlines are requested. This heuristic is not ideal, but stops the program from hanging in some use cases.
Also provide explicit single-thread code when the number of threads is set to 1.