Explicit single-thread handling within multi-threading convenience functions
Added explicit code branching to the convenience functions provided in lib/thread/queue.h, for when the number of threads is set to 1.
Previously, in such a use case, the function would spawn 1 thread for each of source / pipe / sink functors.
Now, if a single thread is requested, the function will create local instances of the templated data type, and call the functors locally, without use of the Thread::Exec class.
Therefore, no additional threads are spawned.
Closes #45