Minor fix to provided scripts
In a rare use case, if the user specified an output image path that matched one of the temporary files generated by the script, it would cause a file overwrite error.
This fix uses a unique name for the final image output within the temporary directory (but using the correct image format), so that it is guaranteed not to conflict with the script, and can still be moved to the user's desired location using a move command rather than mrconvert.