New command shbasis
Command shbasis examines the power present in each SH coefficient volume, and uses the difference in power between m==0 and m!=0 coefficients to estimate whether or not an SH image was generated using the old non-orthonormal basis, or the new orthonormal basis.
Additionally, command-line options may be used to convert the image data in-place to conform to the desired basis.
Discussed in Issue #38.