Thread::Queue: clean up API to make it simpler to use
There is now a single Thread::run_queue() function, with the following
syntax:

    Thread::run_queue (source, item, sink, [capacity]);
    Thread::run_queue (source, item, pipe, item2, sink, [capacity]);

The optional capacity refers to the capacity of the queue(s) in the
pipeline, and default to MRTRIX_QUEUE_DEFAULT_CAPACITY (currently 128).

Each of the source/pipe/queue functors can be specified as being
multi-threaded by wrapping it in a call to Thread::multi(). For example,
this replaces Thread::run_queue_threaded_sink():

    Thread::run_queue (source, item, Thread::multi (sink));

Any item entry can be specified as batched by wrapping it in a call to
Thread::batch(). For example, this replaces
Thread::run_batched_queue_threaded_pipe():

    Thread::run_queue (
          source,
          Thread::batch (item),
          Thread::multi (pipe),
          Thread::batch (item2),
          sink);

All of these features can be used in any combination.

By default, batches will consist of MRTRIX_QUEUE_DEFAULT_BATCH_SIZE
items (currently 128). This can be changed by adding the number as an
extra argument:

    Thread::run_queue (source, Thread::batch (item, 1024), sink);

Similarly, each functor will use the number of threads as given by
Thread::number_of_threads(). If needed, this can be set explicitly if
needed with an additional argument:

    Thread::run_queue (Thread::multi (source, 4), item, sink);