Configure: include PATH in config file, use in build
The PATH environment variable used during ./configure is now written
into the config file. The build script then uses that PATH in the
environment of any commands it executes. This ensures that commands
necessary at compile time are available without having to set the PATH
explicitly every time (this is particularly useful for Qt's moc, which
may be in a non-standard location).