Updates to shbasis
shbasis now only uses the ratio of m==0 to m!=0 power in l=4 to estimate the SH basis.
Furthermore, a check has been added that compares this ratio in l=4 to the ratio in lmax; a decrease in this ratio in higher orders appears to be indicative of a poor encoding scheme.