Image::ThreadedLoop: new run_foreach_pos() methods
These are identical in usage to the current run_foreach() methods, but
expect a function/functor that accepts a const Image::Iterator& as the
first argument. This will contain the position of the voxel currently
being processed, which may be useful in a number of different contexts,
and should make these methods applicable to a broader range of problems.