Updated configure to check for QT resource compiler
The configure script will now check for the QT resource compiler (rcc) in the path.
If it does not have the default name or is not in the path, the RCC environment variable can be set to specify its location.
This behaviour is similar to the handling of other QT tools by the configure script.