FMLS::FODQueueWriter operates on VoxelType rather than BufferType
Creating an instance of a VoxelType within the operator() method of a functor
is to be avoided, since it adds unnecessary work (admittedly probably not
much), and creates a huge amount of debug-level messages (one per invocation),
making debugging very messy.

In general, it is best to always design functors and other generic classes to
operate on VoxelTypes rather than BufferTypes, since this allows them to be
trivially copy-constructed if need be, and more importantly allows the code to
accept non-standard VoxelTypes that have no corresponding BufferType (e.g.
Voxel::Adapter classes).