Implemented fail-on-warn capability
The WARN() macro is typically used to warn the user that their data is somehow incomplete, or some assumption has been made regarding how the processing should proceed.
In batch procesing contexts, it may be preferable to abort processing if such an event occurs.
This can now be achieved using the -failonwarn command-line option, or the FailOnWarn config file entry.
Closes #37.