Image::Info: added functionality to prevent modification of transform
The Image::Info::sanitise_strides() function (invoked by
Image::Info::sanitise(), in turn invoked by the
Image::Header::open()/create() calls) can modify the transform when it
represents a significant deviation from the MRtrix coordinate system
(i.e. it represents a rotation onto the sagittal or coronal planes).
When this happens, the columns of the rotation part of the transform are
exchanged and if required their sign flipped, with corresponding changes
to the strides and a shift in the origin if appropriate.

This is normally the desired behaviour, so an axial slice is always
axial no matter how the data were acquired. There are however cases
where this introduces problems, notably in mrtransform when the data are
rotated such that this process is trigerred. In this case, the
redefinition of the axes means that the copy can fail or lead to odd
behaviour.

The new Image::ProtectTransform class can be used to get around this
problem. When instanciated, this class will set a global variable
instructing the Image::Info::sanitise_strides() to leave the transform
as-is. The global variable is reset to default when the class goes out
of scope (or its stop() method is invoked).