tckgen: set default number of tracks to 1000
Having a default behaviour of doing nothing with no explanation is not a
good idea. I just spent the last half hour hunting down what I thought
must be a bug in the code, only to realise that the default number of
tracks is zero... An application should do something sensible even if no
(non-mandatory) options have been set. Alternative is to force -number
to be mandatory, but that would just get annoying for the many cases
when one just wants to run a quick check.