Simple syntax to document config file options
Configuration file options can now be documented and extracted from the
code using a simple syntax (as suggested in #51). Any line that starts
with //CONF (ignoring white space) is considered as part of that
documentation. The syntax expects one line to document the name of the
option, the next its default value, and the following lines to provide a
description, up to the next option name (which might be in a separate
file altogether). For example:

    //CONF option: NumberOfThreads
    //CONF default: 1
    //CONF set the default number of CPU threads to use for multi-threading.

Documented options can be then be queried using a command such as (this
essentially just matches those lines and strips out the leading //CONF):

    $ grep -rn --include=\*.h --include=\*.cpp '^\s*//CONF\b ' . | sed -ne 's/^.*CONF \(.*\)/\1/p'

Eventually, the output of this could be used to generate the
documentation directly (probably with a further sed command).