Implemented greyscale gradient output for mrfilter
By default, the gradient filter in mrfiler produces a 3D volume for every volume input, with RGB components corresponding to the three image / scanner spatial axes.
Now, if the -greyscale option is added, the output image will have the same dimensions as the input image, and the gradient orientation information is collapsed to a single greyscale intensity.