Added FFT filter to mrfilter
Re-arranged existing contents of lib/image/filter/fft.h to conform to the Image::Filter::Base class.
Class has capability to perform FFT along multiple axes.
Class may also optionally re-arrange data so that the zero-frequency term appears in the centre of the image (better for visualisation).
Although this is made available in the mrfilter command, a separate code path is required due to the use of complex types.
The output image is complex by default, or a greyscale image containing the magnitude of the result is obtained using the -magnitude option.