Variable maximum number of tracking attempts from seed point
As discussed in Issue #41.
Once an appropriate seed point is determined, a probabilistic tracking algorithm may take some number of attempts to find a suitable direction to commence tracking.
The maximum number of attempts now depends on the mechanism used to generate the seed points.
For instance, if the seed points are drawn randomly from a mask image, only 1 direction is tested per seed point; whereas if seeds are defined on a lattice grid, every attempt is made to find a valid tracking direction from every seed point.
This behaviour can be overridden using the -max_seed_attempts option.