Added -stop option to tckgen
Closes #40.
Just as in MRtrix 0.2, the new -stop option instructs the tractography algorithm to cease streamline propagation as soon as a given streamline has entered all inclusion regions.
A warning will be produced if the user tries to apply this option but does not provide any inclusion regions.
This option cannot be used in conjunction with back-tracking in ACT.