tckmap: Strip template header comments
Comments from an image used as a template for producing a TWI may not be relevant to the TWI itself, and only lead to a bloated header (especially if one TWI is used as the template for another).
This change strips away any comments read from the header of the template image.