Clear sparse image intensities on image creation
Previously on sparse image creation, the image intensity field was left un-initialised, with the initialisation left to the coder; this could result in buffer over-runs if the data were not explicitly zeroed for every voxel.
This commit zeroes all image data on formation of a new sparse image, simplifying the implementation of applications using this format.
Closes #53.