Do not pre-load streamline weights
In the initial implementation, streamlines weights from a text file weere imported into a Math::Vector.
This caused an unnerving pause at the start of command execution, and took up unnecessary RAM.
This new implementation reads the weights one at a time as the streamlines data is read from file.
Checks are still made regarding the number of streamlines v.s. the number of weights, but the warnings now occur when the read process is completed.
Closes #37.