MRView: add config file option to force OpenGL3.3 core profile
By default, the GUI backend will only check that the OpenGL version is
sufficient to run MRView (and shview), and issue a warning if not (note
that it doesn't actually abort at this stage, this typically happens due
to an overt error when compiling the shaders, which explicitly require
version 3.3 of the GLSL language). The backend doesn't explicitly
require a core 3.3 profile on most platforms since the default profile
is typically the full compatibility profile, with version greater than
3.3. With NVIDIA drivers on Linux, requesting a core 3.3 profile
actually causes issues with certain primitives, due to what must be a
driver bug - this doesn't happen when using the default profile.

The exception to this is MacOSX, which requires the core profile to run.
However, the latest Mesa drivers on Linux (as of version 10.1) now
include full 3.3 support, particularly the nouveau drivers for nvidia.
To get them to work also requires explicitly requesting a core 3.3
profile. It is therefore no longer sufficient to assume that the
compatibility profile is sufficient on Windows & Linux. To get around
this, a new config file option is available to request (or not) a core
3.3 profile if required.  The default remains as before (core 3.3 on
MacOSX, compatibility on other platforms).