DICOM: do not modify DW information
Remove scaling of b-value by square of norm of DW gradients from DICOM
handling code, and move it to a dedicated
DWI::scale_bvalue_by_G_squared() function, as discussed in issue #48.
This means applications are guaranteed to be provided with an accurate
representation of the DW information in the DICOM headers. The
DWI::get_valid_DW_scheme() function now handles the scaling of the
b-values, but only if the "-scale_bvalue_by_grad" option is also
supplied.

The DWI::get_DW_scheme() function now also returns the unmodified DW
gradient scheme - previously the gradients were normalised before being
returned.