DW scheme handling: add -bvalue_scaling option
As discussed in issue #60, multi-shell schemes will typically specify a
single nominal b-value and modulate according to the amplitude of the DW
gradients. This commit adds an option (and corresponding config file
entry) to specify whether the b-values should be scaled. The default is
to let MRtrix auto-detect whether scaling is needed based on the
presence of more than 2 b-values, which would indicate a multi-shell
scheme.