DICOM: get VR from dict if unknown
Even using explicit transfer syntax, some vendors can't be bothered to
properly fill in the VR (value representation) field, leaving it as UN -
unknown. This prevents reading tags that would otherwise be
straightforward to read. This commit adds a further attempt at figuring
out the VR for items that remain unknown, by looking it up in the DICOM
dictionary - a step that should only be needed for implicit transfer
syntax.