DICOM: don't rely SeriesTime being constant over a DICOM series
This is unfortunately not generally true. Particular example is a GE
dataset that won't have a constant SeriesTime after on-scanner bias
field correction, when the original data set does.

Unfortunately there is no guarantee this won't cause issues with other
data sets. We'll have to wait and see if anyone reports any problems...