New iteratively-reweighted constrained least-squares solver
Basically identical to that currently used in CSD, with regularisation
factor increased upon convergence until a suitable threshold is reached.
Converges in ~30-70 iterations depending on complexity of problem, and
faster than MatLab's quadprog by a factor of ~10. Requires careful
tuning of initial regularisation factor and increment, with optimal
values depending on how well-defined the problem is.