New script revpe_dwicombine
This script is designed to handle cases where a particular DWI acquisition has been repeated twice, with reversal of the phase encode direction on the second acquisition.
THerefore, the script extracts the b=0 images from these volume series to estimate the inhomogeneity field, and produces a single output volume.
The output image should have the same number of volumes as one of the input series; every volume from each of the two image series is combined to produce a single output volume.

Note however that there appears to be a bug in FSL (at least in the versions I have tried) that prevents the script from working; more details in the script file itself.