New QT widget class ComboBoxWithErrorMsg
This new class derives from QComboBox.
When no item in the combo box adequately describes the current selection (e.g. multiple input files are selected, and each has a different value for that parameter), this class appends an extra item to the end of the drop-down list (with a custom string); this extra item is then deleted once a valid selection is made (either by code or user).
This is used in the fixel tool to indicate when the colouring / length-scaling properties are different between the currently selected files.