mrview: Fix fixel tool threshold tristate handling
Previously, if multiple input files were selected, and the state of a threshold varied between the selected files, the check box would become tristate, which would later cause unwanted behaviour.
This change allows for a tristate in this case, but explicitly reverts to boolean in all other cases, and threshold status of input files is preserved through multi-file selection unless the user activates the check box.