Changed deadzone behaviour of AdjustButton class
Previously, an initial vertical movement of more than 4 pixels would engage the mouse-dragging behaviour until the mouse button was released.
Now, this threshold has been doubled to 8 pixels; and if the mouse returns back to within this distance of the click point, the button value will return to the initial value from when the mouse button was first clicked.