afdconnectivity: New capabilities
Previously, afdconnectivity would select fixels based on mapping a track file corresponding to a bundle of interest. This may perform poorly in regions of partial volume, or where erroneous streamlines take a different path between the target regions of interest.
The command now has the capability to first read a whole-brain fibre-tracking data set; by doing so, for each fixel traversed by the bundle of interest, only some fraction of the AFD in that fixel is contributed to the result; this fraction is determined by the streamlines density provided by the bundle of interest, divided by the streamlines density in that fixel in the whole-brain tractogram.
Therefore if only a small part of the fixel corresponds to the bundle of interest, only a small part of its AFD contributes.