Tractography::Mapping: Redesign MapWriter class
Due to excessive code duplication, have cancelled the idea of having a MapWriter type for each output image dimensionality, and instead everything is handled using a single MapWriter class.
An enum tells MapWriterBase the type of writer instantiated, and the SetVoxel class type fed to the functor controls how the image buffer(s) is(are) updated.