Tractography::Mapping: Simplify TrackMapperBase functions
This class now has three voxelise functions: one for straight TDI, one for any case where the streamline tangent is required, and one for the precise mapping method.
The latter two both make use of the add_to_set() functions to calculate the relevant information based on the streamline tangent and contribute the result to the set of elements traversed by the streamline.