Tractography::Mapping: Alterations to SetVoxel classes
Fixed some bugs in the SetVoxel classes, made the function design a little safer.
In the mapping code itself, used Point<int> throughout mapping, which is only converted to a Voxel class when preparing to write to a set; this makes it clearer what represents what where, especially since the Voxel class now contains a length.
Precise mapping appears to be working now.
Also changed the -colour option in tckmap to -dec.