tckmap: Clean up output image header
Previously properties from the track file, as well as any template image, were propagated through to the output image.
This typically made the header information huge, and would take up a lot of space in mrview.
Instead, now just a few properties related to the TWI process are written to the header.