tckmap: Re-implemented gaussian track-wise statistic
Note that tckmap functionality is untested; this is simply the first version that compiled.
The Gaussian track-wise statistic stands out because it necessitates an independent TWI 'factor' per element that the streamline contributes to, rather than a single value for the whole streamline.
To better accommodate this in the presence of more complex output image dimensionality, a separate Mapper class (in src/dwi/tractography/mapping/gaussian/mapper.h) has been provided that performs the additional work during the voxelise() methods, and separate SetVoxel classes (in src/dwi/tractography/mapping/gaussian/voxel.h) that include the additional necessary information per mapped element.
The MapWriter class receives either a standard Set or one of the sets as defined in the Gaussian namespace, and reads the factor per mapped element from either the mapped element itself or the SetVoxelExtras class accordingly.