MRView: update screen after all pending events have been processed
This addresses issue #80, which was concerned with some serious lag due
to mouse events being processed one by one combined with slow (or even
not so slow) OpenGL refreshes. In this commit, all updateGL() calls are
now fed through to a single-shot QTimer connected to the GLArea's actual
updateGL() call. The timer is set with a zero timeout, which according
to the Qt documentation, guarantees it will be called once all events
have been processed.