GMWMI seeding: Change to floating-point image
Command 5tt2gmwmi will now generate a floating-point image for GMWMI seeding, rather than a mask.
Internally, the GMWMI seeder then uses the rejection seeder to determine the initial seed points before interface optimisation.
Using tri-linear interpolation in the rejection seeder prior to seed point selection, rather than after, was implemented, but found to not have a significant effect in this circumstance, so is not used by default (as it will otherwise result in seeds in voxels with zero intensity adjacent to non-zero intensity voxels)