Major changes to handle thread-safe access to bitfields
Now use std::atomic functionality to allow thread-safe access to
bit-wise data, as per issue #98. **This means MRtrix now requires
the compiler to support C++11.** This is also relevant for issues #32
(move to C++11) and #31 (switch to non-GPL license).

The use of the std:atomic::compare_exchange_weak() function is also a
bit of a hack, relying on the implementation of std::atomic<T> being a
thin wrapper around its type T, with no additional state. This is the
case with GCC on Linux, and I expect any other compiler on relatively
recent Intel-compatible hardware. While this should cover the
overwhelming majority of likely MRtrix installations, there is a
distinct possibility that this will fail to compile or run on different
CPUs.

Will increment version number to duly reflect the ominousness of the
occasion...