Change order of arguments to Image::ThreadedLoop
This changes the order of arguments for the two constructors that take
min & max axes to loop over, and moves the num_inner_axes argument to
last with default 1 (which is what is used in the overwhelming majority
of cases). This makes more sense from a developer point of view. Means
typical calls will now change from:

    Image::ThreadedLoop loop (vox, 1, 0, 3).run (...);

to

    Image::ThreadedLoop loop (vox, 0, 3).run (...);

and where required, the number of inner axes can be added at the end:

    Image::ThreadedLoop loop (vox, 0, 3, 2).run (...);