New command distdir to distribute directions over subsets
This is used to split a previously generated set of directions as
generated using gendir into subsets, each to be acquired with a
different phase encoding direction. It also optimise these directions to
sample the space of eddy-current distortions as uniformly as possible,
as required by FSL5's EDDY, by flipping the sign of directions.