mrview: Reduce lag from ODF render frame
The DWI render frame was calling QGLWidget::updateGL() directly, which could cause severe lag when moving the focus point (even with ODF overlay turned off).
Used the same trick as that used to solve Issue #80.
Hopefully solves the remnant of Issue #114, but leaving it open for a little while because I'm not sure whether or not this is the only remaining source of input lag.