New command 'splicedir' to merge DW_schemes
This assumes a number of subsets per b-value each corresponding to a
different PE direction, and merges them into one large encoding file
with their order randomly reshuffled. The resulting file is a 5 column
format, with the (normalised) X,Y,Z gradient direction, b-value, and an
index corresponding to the PE direction.