Image::Stride::sanitise() fix
Fix the case where a limited constraint on the image strides has been provided, and the remaining strides are to be determined based on a reference image.
This was otherwise causing an error when using Image::Stride::contiguous_along_axis(3) if the template InfoType had been expanded from 3D to 4D (and hence stride[3]=0).
Fixes #142.