mrview: Fix ROI tool draw / erase buttons
Previously, whenever either the draw or erase button was checked, and the other button was then clicked (thereby automatically unchecking the other), the first button would then misbehave on the next button press.
It seems as though there is some latent variable in the QToolButton class that is not being correctly modified by the setChecked() function, such that the next user click does the wrong thing.
Only way I could fix it was by calling the click() function for the button to be unchecked (even toggle() didn't fix the problem).