mrview ROI tool: Improved single-voxel draw behaviour
If the brush is being used in the ROI tool, and the brush size is set to the minimum permissible, the tool will call new function draw_line().
This uses a ray-tracing algorithm to highlight all voxels intersected by the line between the current mouse position and the previous position.
This prevents holes from appearing in the path drawn if the mouse is moved quickly.