dwi2response: Meaningful message for rare error
In some data where an erroneous gradient scheme was used, the initial A2SH transformation for computing FODs works, but once the gradient table is rotated according to the peak FOD orientation for estimating m=0 coefficients of the response function, the A2SH transformation can no longer be derived.
This changes catches the error and gives the user a meaningful warning, rather than having the command terminate with an ambiguous message.
Discussed in #167.