Fix handling of exceptions in threads
Previous commit introduced possibility of throwing in destructor of
thread object, which is not recommended (to the extent that destructors
in C++11 are by default not allowed to throw). This introduces a new
wait() method on the thread objects that can throw if any of the
corresponding threads do throw an Exception.