DWI::Shells: bzero_threshold no longer a global variable
The bzero_threshold variable used in the DWI::Shells class was declared
as a global variable, and initialised from File::Config. Problem is that
initialisation for a global variable happens before entering main(), and
hence before Config::init() (in App::parse()). So there is no way that
the Config::get() call could return anything other than the default
supplied, since the Config would appear empty at that point...

Solution proposed here is to  return a static member variable in a
function, which defers initialisation until the first time it's called.
Hopefully that function will never be called before Config::init()...