Fix stderr redirection bug on Windows GUI apps
Seems the GUI commands (mrview, shview) don't get console access on Windows, so the test implemented in 1530151 was giving a 'bad file descriptor' error.
This fix just catches the resulting exception and continues sensibly.
Note however that it's impossible to pipe stderr to a file for GUI applications on Windows.
Closes #197.