Add rudimentary testing framework
We need something like this for regression testing, especially ahead of
potentially large changes in the codebase if/when we move to Eigen.

So far, this adds a couple of commands to generate and diff test data, a
set of scripts to execute for the tests in the testing/ folder, and a
run_tests script to run all these scripts in turn. Nothing fancy at this
stage, the idea really is to generate some test data and process it
using a known-good version of the code, then save both inputs and
outputs in the testing/data/ folder. In the scripts, we run the same
commands and check against the saved data for consistency.

Not sure how to handle stochastic processes at this stage - maybe this
will only be possible by disabling multi-threading and supplying an
explicit seed for the random number generator.