MRtrix GZ format: Fix header size
Issue #207 may have arisen because the GZ handler needs to store header information in the internal lead-in member data; it does this by first calling read_mrtrix_header() to fill an Image::Header, then calling write_mrtrix_header() to generate a new header string.
With the change implemented in #111, the input and output header strings are no longer guaranteed to be of the same legnth (particularly as without the fix implemented in 24c0a61, the header was extending considerably).
This fix now separately tracks the offset of the image data from the start of the input file, and the size of the output header text / GZ lead-in, which will hopefuilly fix the problem for good.