Connectome tool: Initial work on flexible node visualisation
Have laid the foundations for allowing a high level of flexibility in choosing how nodes of the connectome are plotted.
Most of these don't actually work yet, I'm just getting the framework in place for selecting them and updating settings accordingly.
I also added the possibility of controlling the transparency of each node, not just its visibility.