syntax overhaul: remove datatype() method from Image
The datatype() method is now reserved for the Header, and refers
specifically to the data type as stored on file. Since this can easily
differ from the types returned by the Image::value() call, having
Image::datatype() return the Header's datatype would probably just lead
to confusion... The datatype of the Image can be queried using its
value_type member typedef, and converted to a Datatype if needed using
the static Datatype::from<value_type>() function.